Relocate vector wrappers
