+++ gymnasium/experimental/wrappers/vector/__init__.py
@@ -1,0 +1,12 @@
+"""Wrappers for vector environments."""
+
+from gymnasium.experimental.wrappers.vector.record_episode_statistics import (
+    VectorRecordEpisodeStatistics,
+)
+from gymnasium.experimental.wrappers.vector.vector_list_info import VectorListInfo
+
+
+__all__ = [
+    "VectorRecordEpisodeStatistics",
+    "VectorListInfo",
+]
